move EventSourcing.ForgettablePayloads.Persistence.SqlServer DI code into main EventSourcing.ForgettablePayloads.Persistence.SqlServer package
