+++ EventSourcing.ForgettablePayloads/Persistence/SqlServer/ForgettablePayloads.Persistence.SqlServer.IntegrationTests/Extensions/EventSourcingForgettablePayloadsBuilderExtensions_Should.cs
@@ -1,13 +1,13 @@
 ﻿using EventSourcing.Extensions;
 using EventSourcing.ForgettablePayloads.Extensions;
-using EventSourcing.ForgettablePayloads.Extensions.DependencyInjection.Persistence.SqlServer;
+using EventSourcing.ForgettablePayloads.Persistence.SqlServer.Extensions;
 using Microsoft.Extensions.DependencyInjection;
 using TestHelpers.Extensions;
 using Xunit;
 
-namespace Extensions.DependencyInjection.SqlServer.UnitTests
+namespace ForgettablePayloads.Persistence.SqlServer.IntegrationTests.Extensions
 {
-    public class ServiceProvider_Should
+    public class EventSourcingForgettablePayloadsBuilderExtensions_Should
     {
         [Fact]
         public void BuildWithoutErrors_When_AddingEventSourcingAndExternalDependenciesAndForgettablePayloadsUsingSeparateBuilders()
