+++ EventSourcing.ForgettablePayloads/Persistence/SqlServer/ForgettablePayloads.Persistence.SqlServer.IntegrationTests/SqlServerCollectionFixture.cs
@@ -6,8 +6,8 @@ using EventSourcing.Extensions;
 using EventSourcing.Extensions.DependencyInjection.Serialization.Json;
 using EventSourcing.ForgettablePayloads.Extensions;
 using EventSourcing.ForgettablePayloads.Extensions.DatabaseMigrations.Persistence.SqlServer.DbUp.Extensions;
-using EventSourcing.ForgettablePayloads.Extensions.DependencyInjection.Persistence.SqlServer;
 using EventSourcing.ForgettablePayloads.Persistence.SqlServer;
+using EventSourcing.ForgettablePayloads.Persistence.SqlServer.Extensions;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
